requirements.txt
streamlit
pandas
python-dotenv
SQLAlchemy>=2.0
psycopg2-binary
openpyxl
reportlab
plotly